// client/src/generated-ts-client.ts
export interface PendingDepositsListItemDto {
    transactionId?: string;
    playerId?: string;
    playerFullName?: string;
    mobilePayReference?: string;
    amountDkk?: number;
    createdAt?: string;
}